adding unsafe gmm version
